+++ extensions/gitlab/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Search Scope] - 2023-04-27
 
 - Add scope dropdown for `Search Issues` and `Search Merge Requests`
-- Add possibility to search for specific `state`, `author`, `assignees`, `milestone` or `label` in `Search Issues`.
+- Add possibility to search for named parameters in `Search Issues` and `Search Merge Requests`.
 
 ## [Remove Branch] - 2023-04-17